Created SkeletonState class with seperate source file. Modified size of renderered skeleton. Hand point activity is reflected by color and size.
