+++ src/vector.h
@@ -23,7 +23,7 @@ template <class T> void vectorSubtraction(const T a[3], const T b[3], T c[3])
     c[0] = a[0] - b[0];
     c[1] = a[1] - b[1];
     c[2] = a[2] - b[2];
-    
+
     return;
 }
